requirements.txt
torch
torchvision
pytest 
pandas
numpy
scikit-learn
matplotlib
seaborn
notebook
dvc
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
scipy
FastAPI
uvicorn
pydantic
requests
python-multipart
streamlit
pathlib
#-e .
ultralytics
